piper-server/pyproject.toml
[project]
name = "piper-server"
version = "0.1.0"
description = "Standalone OpenAI-compatible Piper TTS server."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "python-multipart>=0.0.9",
  "pyyaml>=6.0.1",
  "huggingface-hub>=0.31.0",
  "numpy>=1.26.0",
  "onnxruntime>=1.18.0",
  "piper-tts>=1.3.0",
]

[project.scripts]
piper-server = "piper_server.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
